Add DeltaStrategyGameControllerTestDouble
This test double speeds up the process of testing by providing a new
method "setFieldConfiguration" that can take either a new board manager
or a collection of pieces. This allows a tester to make the test double
with a valid configuration then immediately set the configuration to be
something more convenient for testing purposes. An example would be to
set the field to have both players only have a marshal and a flag to
quickly test end game scenarios.